requirements.txt
opencv-python
Pillow
python-ffmpeg
numpy
scipy
soundfile
ffmpeg-python